requirements.txt
Flask
flask-bcrypt
flask-login
flask-cors
pymongo
PyJWT
fpdf
pytz
num2words
razorpay